MOD: updated a lot dpos api related. moved eth public dpos api to ethapi
