+++ consensus/dpos/api.go
@@ -22,6 +22,29 @@ type API struct {
 	dpos  *Dpos
 }
 
+// CandidateInfo stores detailed candidate information
+type CandidateInfo struct {
+	Candidate   common.Address `json:"candidate"`
+	Deposit     common.BigInt  `json:"deposit"`
+	Votes       common.BigInt  `json:"votes"`
+	RewardRatio uint64         `json:"reward_distribution"`
+}
+
+// ValidatorInfo stores detailed validator information
+type ValidatorInfo struct {
+	Validator   common.Address `json:"validator"`
+	Votes       common.BigInt  `json:"votes"`
+	EpochID     int64          `json:"current_epoch"`
+	MinedBlocks int64          `json:"epoch_mined_blocks"`
+	RewardRatio uint64         `json:"reward_distribution"`
+}
+
+// DelegatorInfo stores detailed DelegatorInfo
+type DelegatorInfo struct {
+	Deposit common.BigInt    `json:"deposit"`
+	Votes   []common.Address `json:"votes"`
+}
+
 // GetConfirmedBlockNumber retrieves the latest irreversible block
 func (api *API) GetConfirmedBlockNumber() (*big.Int, error) {
 	var err error
@@ -43,6 +66,27 @@ func (api *API) GetConfirmedBlockNumber() (*big.Int, error) {
 	return header.Number, nil
 }
 
+// GetValidatorInfo will return the detailed validator information
+func GetValidatorInfo(stateDb *state.StateDB, dposCtx *types.DposContext, header *types.Header, address common.Address) (ValidatorInfo, error) {
+	if err := checkIsValidator(dposCtx, address); err != nil {
+		return ValidatorInfo{}, err
+	}
+	votes := GetTotalVote(stateDb, address)
+	rewardRatio := GetRewardRatioNumeratorLastEpoch(stateDb, address)
+	epochID := CalculateEpochID(header.Time.Int64())
+	minedCount, err := dposCtx.GetMinedCnt(epochID, address)
+	if err != nil {
+		return ValidatorInfo{}, err
+	}
+	// return validator information
+	return ValidatorInfo{
+		Votes:       votes,
+		RewardRatio: rewardRatio,
+		MinedBlocks: minedCount,
+		EpochID:     epochID,
+	}, nil
+}
+
 // checkIsValidator checks if the given address is a validator address
 func checkIsValidator(dposCtx *types.DposContext, addr common.Address) error {
 	validators, err := dposCtx.GetValidators()
@@ -58,35 +102,43 @@ func checkIsValidator(dposCtx *types.DposContext, addr common.Address) error {
 	return fmt.Errorf("the given address %s is not a validator's address", addr.String())
 }
 
-// GetValidatorInfo will return the detailed validator information
-func GetValidatorInfo(stateDb *state.StateDB, dposCtx *types.DposContext, validatorAddress common.Address, header *types.Header) (common.BigInt, uint64, int64, int64, error) {
-	if err := checkIsValidator(dposCtx, validatorAddress); err != nil {
-		return common.BigInt0, 0, 0, 0, err
-	}
-	votes := GetTotalVote(stateDb, validatorAddress)
-	rewardRatio := GetRewardRatioNumeratorLastEpoch(stateDb, validatorAddress)
-	epochID := CalculateEpochID(header.Time.Int64())
-	minedCount, err := dposCtx.GetMinedCnt(epochID, validatorAddress)
+// GetCandidateInfo returns the detailed candidates information
+func GetCandidateInfo(stateDb *state.StateDB, dposCtx *types.DposContext, address common.Address) (CandidateInfo, error) {
+	isCand, err := IsCandidate(dposCtx, address)
 	if err != nil {
-		return common.BigInt0, 0, 0, 0, err
-	}
-	// return validator information
-	return votes, rewardRatio, minedCount, epochID, nil
-}
-
-// GetCandidateInfo will return the detailed candidates information
-func GetCandidateInfo(stateDb *state.StateDB, dposCtx *types.DposContext, candidateAddress common.Address, header *types.Header) (common.BigInt, common.BigInt, uint64, error) {
-	isCand, err := IsCandidate(dposCtx, candidateAddress)
-	if err != nil {
-		return common.BigInt0, common.BigInt0, 0, err
+		return CandidateInfo{}, err
 	}
 	if !isCand {
-		return common.BigInt0, common.BigInt0, 0, fmt.Errorf("address %x is not a candidate", candidateAddress)
+		return CandidateInfo{}, fmt.Errorf("address %x is not a candidate", address)
 	}
 	// get detailed candidates information
-	candidateDeposit := GetCandidateDeposit(stateDb, candidateAddress)
-	candidateVotes := CalcCandidateTotalVotes(candidateAddress, stateDb, dposCtx.CandidateTrie())
-	rewardRatio := GetRewardRatioNumerator(stateDb, candidateAddress)
+	candidateDeposit := GetCandidateDeposit(stateDb, address)
+	candidateVotes := CalcCandidateTotalVotes(address, stateDb, dposCtx.CandidateTrie())
+	rewardRatio := GetRewardRatioNumerator(stateDb, address)
 
-	return candidateDeposit, candidateVotes, rewardRatio, nil
+	return CandidateInfo{
+		Deposit:     candidateDeposit,
+		Votes:       candidateVotes,
+		RewardRatio: rewardRatio,
+	}, nil
+}
+
+// GetDelegatorInfo returns the detailed delegator information
+func GetDelegatorInfo(stateDb *state.StateDB, dposCtx *types.DposContext, address common.Address) (DelegatorInfo, error) {
+	isDelegator, err := HasVoted(dposCtx, address)
+	if err != nil {
+		return DelegatorInfo{}, err
+	}
+	if !isDelegator {
+		return DelegatorInfo{}, fmt.Errorf("address %x is not a delegator", address)
+	}
+	deposit := GetVoteDeposit(stateDb, address)
+	votes, err := dposCtx.GetVotedCandidatesByAddress(address)
+	if err != nil {
+		return DelegatorInfo{}, err
+	}
+	return DelegatorInfo{
+		Deposit: deposit,
+		Votes:   votes,
+	}, nil
 }
