+++ internal/ethapi/dpos_api.go
@@ -6,190 +6,92 @@ package ethapi
 
 import (
 	"context"
-	"errors"
-	"fmt"
 
 	"github.com/DxChainNetwork/godx/common"
 	"github.com/DxChainNetwork/godx/consensus/dpos"
-	"github.com/DxChainNetwork/godx/core/vm"
+	"github.com/DxChainNetwork/godx/core/state"
+	"github.com/DxChainNetwork/godx/core/types"
 	"github.com/DxChainNetwork/godx/rpc"
 )
 
-// PublicDposTxAPI exposes the dpos tx methods for the RPC interface
-type PublicDposTxAPI struct {
-	b         Backend
-	nonceLock *AddrLocker
+// PublicDposAPI is the public dpos api
+type PublicDposAPI struct {
+	b Backend
 }
 
-// CandidateInfo stores detailed candidate information
-type CandidateInfo struct {
-	Candidate   common.Address `json:"candidate"`
-	Deposit     common.BigInt  `json:"deposit"`
-	Votes       common.BigInt  `json:"votes"`
-	RewardRatio uint64         `json:"reward_distribution"`
+// NewPublicDposAPI creates a PublicDposAPI object that is used
+// to access all DPOS API Method
+func NewPublicDposAPI(b Backend) *PublicDposAPI {
+	return &PublicDposAPI{
+		b: b,
+	}
 }
 
-// ValidatorInfo stores detailed validator information
-type ValidatorInfo struct {
-	Validator   common.Address `json:"validator"`
-	Votes       common.BigInt  `json:"votes"`
-	EpochID     int64          `json:"current_epoch"`
-	MinedBlocks int64          `json:"epoch_mined_blocks"`
-	RewardRatio uint64         `json:"reward_distribution"`
-}
-
-// NewPublicDposTxAPI construct a PublicDposTxAPI object
-func NewPublicDposTxAPI(b Backend, nonceLock *AddrLocker) *PublicDposTxAPI {
-	return &PublicDposTxAPI{b, nonceLock}
-}
-
-// SendApplyCandidateTx submit a apply candidate tx.
-// the parameter ratio is the award distribution ratio that candidate state.
-func (pd *PublicDposTxAPI) SendApplyCandidateTx(fields map[string]string) (common.Hash, error) {
-	to := vm.ApplyCandidateContractAddress
-	ctx := context.Background()
-
-	stateDB, _, err := pd.b.StateAndHeaderByNumber(ctx, rpc.LatestBlockNumber)
-	if err != nil {
-		return common.Hash{}, err
-	}
-
-	// parse precompile contract tx args
-	args, err := ParseAndValidateCandidateApplyTxArgs(to, DposTxGas, fields, stateDB, pd.b.AccountManager())
-	if err != nil {
-		return common.Hash{}, err
-	}
-
-	txHash, err := sendPrecompiledContractTx(ctx, pd.b, pd.nonceLock, args)
-	if err != nil {
-		return common.Hash{}, err
-	}
-	return txHash, nil
-}
-
-// SendCancelCandidateTx submit a cancel candidate tx
-func (pd *PublicDposTxAPI) SendCancelCandidateTx(from common.Address) (common.Hash, error) {
-	to := vm.CancelCandidateContractAddress
-	ctx := context.Background()
-
-	if err := pd.validateCancelCandidateRequest(ctx, from); err != nil {
-		return common.Hash{}, err
-	}
-	// construct args
-	args := NewPrecompiledContractTxArgs(from, to, nil, nil, DposTxGas)
-
-	// send contract transaction
-	txHash, err := sendPrecompiledContractTx(ctx, pd.b, pd.nonceLock, args)
-	if err != nil {
-		return common.Hash{}, err
-	}
-	return txHash, nil
-}
-
-func (pd *PublicDposTxAPI) validateCancelCandidateRequest(ctx context.Context, addr common.Address) error {
-	_, dposCtx, _, err := pd.b.StateDposCtxAndHeaderByNumber(ctx, rpc.LatestBlockNumber)
-	if err != nil {
-		return err
-	}
-	if dposCtx == nil {
-		return errors.New("unknown block number")
-	}
-	isCand, err := dpos.IsCandidate(dposCtx, addr)
-	if err != nil {
-		return err
-	}
-	if !isCand {
-		return ErrNotCandidate
-	}
-	return nil
-}
-
-// SendVoteTx submit a vote tx
-func (pd *PublicDposTxAPI) SendVoteTx(fields map[string]string) (common.Hash, error) {
-	to := vm.VoteContractAddress
-	ctx := context.Background()
-
-	stateDB, _, err := pd.b.StateAndHeaderByNumber(ctx, rpc.LatestBlockNumber)
-	if err != nil {
-		return common.Hash{}, err
-	}
-
-	// parse precompile contract tx args
-	args, err := ParseAndValidateVoteTxArgs(to, DposTxGas, fields, stateDB, pd.b.AccountManager())
-	if err != nil {
-		return common.Hash{}, err
-	}
-
-	txHash, err := sendPrecompiledContractTx(ctx, pd.b, pd.nonceLock, args)
-	if err != nil {
-		return common.Hash{}, err
-	}
-	return txHash, nil
-}
-
-// SendCancelVoteTx submit a cancel vote tx
-func (pd *PublicDposTxAPI) SendCancelVoteTx(from common.Address) (common.Hash, error) {
-	to := vm.CancelVoteContractAddress
-	ctx := context.Background()
-
-	if err := pd.validateCancelVoteRequest(ctx, from); err != nil {
-		return common.Hash{}, err
-	}
-	// construct args
-	args := NewPrecompiledContractTxArgs(from, to, nil, nil, DposTxGas)
-
-	// send the contract transaction
-	txHash, err := sendPrecompiledContractTx(ctx, pd.b, pd.nonceLock, args)
-	if err != nil {
-		return common.Hash{}, err
-	}
-	return txHash, nil
-}
-
-func (pd *PublicDposTxAPI) validateCancelVoteRequest(ctx context.Context, addr common.Address) error {
-	_, dposCtx, _, err := pd.b.StateDposCtxAndHeaderByNumber(ctx, rpc.LatestBlockNumber)
-	if err != nil {
-		return err
-	}
-	if dposCtx == nil {
-		return ErrUnknownBlockNumber
-	}
-	hasVoted, err := dpos.HasVoted(dposCtx, addr)
-	if err != nil {
-		return err
-	}
-	if !hasVoted {
-		return fmt.Errorf("failed to send cancel vote transaction, %x has not voted before", addr)
-	}
-	return nil
-}
-
-// Validators will return a list of validators based on the blockNumber provided
-func (pd *PublicDposTxAPI) Validators(blockNr *rpc.BlockNumber) ([]common.Address, error) {
-	num := blockNumberConverter(blockNr)
-	ctx := context.Background()
-	_, dposCtx, _, err := pd.b.StateDposCtxAndHeaderByNumber(ctx, num)
+// Validators returns a list of validators based on the blockNumber provided
+func (d *PublicDposAPI) Validators(blockNr *rpc.BlockNumber) ([]common.Address, error) {
+	_, dposCtx, _, err := d.stateDposCtxAndHeaderByNumber(blockNr)
 	if err != nil {
 		return []common.Address{}, err
-	}
-	if dposCtx == nil {
-		return []common.Address{}, ErrUnknownBlockNumber
 	}
 	return dposCtx.GetValidators()
 }
 
-// Validator return a detailed info about the validator address provided.
-func (pd *PublicDposTxAPI) Validator(address common.Address, blockNr *rpc.BlockNumber) (ValidatorInfo, error) {
+// Validator returns a detailed info about the validator address provided.
+func (d *PublicDposAPI) Validator(address common.Address, blockNr *rpc.BlockNumber) (dpos.ValidatorInfo, error) {
+	statedb, dposCtx, header, err := d.stateDposCtxAndHeaderByNumber(blockNr)
+	if err != nil {
+		return dpos.ValidatorInfo{}, err
+	}
+	return dpos.GetValidatorInfo(statedb, dposCtx, header, address)
+}
+
+// Candidates returns a list of candidates information based on the blockNumber provided
+func (d *PublicDposAPI) Candidates(blockNr *rpc.BlockNumber) ([]common.Address, error) {
+	_, dposCtx, _, err := d.stateDposCtxAndHeaderByNumber(blockNr)
+	if err != nil {
+		return []common.Address{}, err
+	}
+	return dposCtx.GetCandidates()
+}
+
+// Candidate returns detailed candidate's information based on the candidate address provided
+func (d *PublicDposAPI) Candidate(address common.Address, blockNr *rpc.BlockNumber) (dpos.CandidateInfo, error) {
+	statedb, dposCtx, _, err := d.stateDposCtxAndHeaderByNumber(blockNr)
+	if err != nil {
+		return dpos.CandidateInfo{}, err
+	}
+	return dpos.GetCandidateInfo(statedb, dposCtx, address)
+}
+
+func (d *PublicDposAPI) Delegator(address common.Address, blockNr *rpc.BlockNumber) (dpos.DelegatorInfo, error) {
+	statedb, dposCtx, _, err := d.stateDposCtxAndHeaderByNumber(blockNr)
+	if err != nil {
+		return dpos.DelegatorInfo{}, err
+	}
+	return dpos.GetDelegatorInfo(statedb, dposCtx, address)
+}
+
+// EpochID  calculates the epoch id based on the block number provided
+func (d *PublicDposAPI) EpochID(blockNr *rpc.BlockNumber) (int64, error) {
+	_, _, header, err := d.stateDposCtxAndHeaderByNumber(blockNr)
+	if err != nil {
+		return 0, nil
+	}
+	return dpos.CalculateEpochID(header.Time.Int64()), nil
+}
+
+// stateDposCtxAndHeaderByNumber is the adapter function for PublicDposAPI.StateDposCtxAndHeaderByNumber
+func (d *PublicDposAPI) stateDposCtxAndHeaderByNumber(blockNr *rpc.BlockNumber) (*state.StateDB, *types.DposContext, *types.Header, error) {
 	num := blockNumberConverter(blockNr)
 	ctx := context.Background()
-	_, dposCtx, statedb, err := pd.b.StateDposCtxAndHeaderByNumber(ctx, num)
+	statedb, dposCtx, header, err := d.b.StateDposCtxAndHeaderByNumber(ctx, num)
 	if err != nil {
-		return ValidatorInfo{}, err
+		return nil, nil, nil, err
 	}
-	if dposCtx == nil || statedb == nil {
-		return ValidatorInfo{}, ErrUnknownBlockNumber
+	if header == nil || statedb == nil || dposCtx == nil {
+		return nil, nil, nil, ErrUnknownBlockNumber
 	}
-
+	return statedb, dposCtx, header, err
 }
 
 func blockNumberConverter(blockNumber *rpc.BlockNumber) rpc.BlockNumber {
